done put

+++ src/main/java/com/example/testorder/repositories/OrdersRepository.java
@@ -6,5 +6,5 @@ import org.springframework.stereotype.Repository;
 
 
 @Repository
-public interface OrdersRepository extends JpaRepository<Order,Integer> {
+public interface OrdersRepository extends JpaRepository<Order,Long> {
 }
